chore: copy protobuf compiler tool from mono repo
This patch copies the cnidarium related parts of
the penumbra protobuf compiler tool [1]. Having
this tool as part of the cnidarium repo is
necessary to update to newer versions of
ibc-types, ibc-proto, prost, tonic, and ics23
because these are also targeted by the code
generator.

This patch contains no logic (such as CI
workflows) to maintain parity with the penumbra
mono repo. This work is left to a follow-up PR.

1: https://github.com/penumbra-zone/penumbra/tree/159f48489bc8deb654551402fdce61108409c524/tools/proto-compiler